+++ review/src/main/java/com/vatunisia/soh/review/repository/ReviewRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    List<Review> findByBusinessId(Long businessId);
+    List<Review> findByBusinessId(Integer businessId);
 }
