+++ review/src/main/java/com/vatunisia/soh/review/service/ReviewService.java
@@ -5,8 +5,8 @@ import com.vatunisia.soh.review.entity.Review;
 import java.util.List;
 
 public interface ReviewService {
-    List<Review> getAllReviews(Long businessId);
-    boolean addReview(Long businessId, Review review);
+    List<Review> getAllReviews(Integer businessId);
+    boolean addReview(Integer businessId, Review review);
     Review getReview(Long reviewId);
     boolean updateReview(Long reviewId, Review review);
     boolean deleteReview(Long reviewId);
